+++ pkg/http/handler.go
@@ -18,18 +18,20 @@ import (
 
 	"github.com/slok/kubewebhook/v2/pkg/log"
 	"github.com/slok/kubewebhook/v2/pkg/model"
+	"github.com/slok/kubewebhook/v2/pkg/tracing"
 	"github.com/slok/kubewebhook/v2/pkg/webhook"
 )
 
 var (
-	runtimeScheme = func() *runtime.Scheme {
+	admissionReviewDeserializer = func() runtime.Decoder {
 		r := runtime.NewScheme()
 		r.AddKnownTypes(admissionv1beta1.SchemeGroupVersion, &admissionv1beta1.AdmissionReview{})
 		r.AddKnownTypes(admissionv1.SchemeGroupVersion, &admissionv1.AdmissionReview{})
-		return r
+
+		codecs := serializer.NewCodecFactory(r)
+
+		return codecs.UniversalDeserializer()
 	}()
-	codecs       = serializer.NewCodecFactory(runtimeScheme)
-	deserializer = codecs.UniversalDeserializer()
 )
 
 // MustHandlerFor it's the same as HandleFor but will panic instead of returning
@@ -46,6 +48,7 @@ func MustHandlerFor(config HandlerConfig) http.Handler {
 type HandlerConfig struct {
 	Webhook webhook.Webhook
 	Logger  log.Logger
+	Tracer  tracing.Tracer
 }
 
 func (c *HandlerConfig) defaults() error {
@@ -58,6 +61,11 @@ func (c *HandlerConfig) defaults() error {
 	}
 	c.Logger = c.Logger.WithValues(log.Kv{"svc": "http.Handler"})
 
+	if c.Tracer == nil {
+		c.Tracer = tracing.Noop
+	}
+	c.Tracer = c.Tracer.WithValues(map[string]interface{}{"svc": "http.Handler"})
+
 	return nil
 }
 
@@ -69,14 +77,19 @@ func HandlerFor(config HandlerConfig) (http.Handler, error) {
 		return nil, fmt.Errorf("handler invalid configuration: %w", err)
 	}
 
-	return handler{
+	h := config.Tracer.TraceHTTPHandler("webhookHTTPHandler", handler{
 		webhook: config.Webhook,
-		logger:  config.Logger}, nil
+		logger:  config.Logger,
+		tracer:  config.Tracer,
+	})
+
+	return h, nil
 }
 
 type handler struct {
 	webhook webhook.Webhook
 	logger  log.Logger
+	tracer  tracing.Tracer
 }
 
 func (h handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
@@ -115,6 +128,7 @@ func (h handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		"ns":           ar.Namespace,
 		"name":         ar.Name,
 		"path":         r.URL.Path,
+		"trace-id":     h.tracer.TraceID(ctx),
 	})
 	logger := h.logger.WithCtxValues(ctx)
 
@@ -187,7 +201,7 @@ func (h handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}).Infof("Admission review request handled")
 }
 func (h handler) requestBodyToModelReview(body []byte) (*model.AdmissionReview, error) {
-	kubeReview, _, err := deserializer.Decode(body, nil, nil)
+	kubeReview, _, err := admissionReviewDeserializer.Decode(body, nil, nil)
 	if err != nil {
 		return nil, fmt.Errorf("could not decode the admission review from the request: %w", err)
 	}